#403: Memory map cleanup - rework memory object
A memory map now contains multiple spans.  The input hex files contain final
addresses which are relocated (if needed) at load time in llrt.

Multiple spans required changing the interface to lower levels of the API which
expect a vector.  This code uses an iterator to the vector.  Ideally,
eventually, the low level code should change to just take a pointer to a slab
of memory.

This commit mostly works but breaks a couple things (ncrisc relocation).