#403: Cleanup remove contextASM/ncrisc reset code
This was used to handle ncrisc transitions from L1 code to iram code on GS
because bbe ncrisc fw overflowed the iram size.  metal doesn't use this at
present and it is a fair amount of complexity and ncrisc L1 code support
will be removed in a future commit.  It can all can be added back after
the new memory map is in place if needed.