#403: Refactor in prep for fw/kernel split
This commit should have no functional impact.

This commit is split out from what was a larger commit with a regression.
Checking in in pieces to make debug easier.
Adds a lot of "attribute((used))" to globals so that after the fw/kernel split
they don't get optimized out of the fw.