#403: Use one L1 loc for brisc/ncrisc/trisc sync
This was the original implementation, but it ran into a bizarre seeming HW bug
where readers were reading different values from the same location at once.
A lot has changed since then and this now seems to work.  The old issue was
never root caused, but was probably a boneheaded SW error.