#5137: Use memcpy for float conversions
For some reason, newer GCC refuses to treat the uint32_data variable as
initialized if reinterpret_cast<> is used. Possibly because on some
architectures the size difference of uint32_t and float means that it
could be partially uninitialized?

Anyway, the other standard pattern in C++ of using memcpy seems to not
hit the same (false) positive, so let's go with that.

Once we're on C++20, std::bit_cast will come in handy here.

Signed-off-by: Olof Johansson <olofj@tenstorrent.com>