#5137: tt_eager: Fix a few possibly uninitialized variable warnings
I'll be honest, I'm not deeply familiar with the C++ rules and why the
previous constructs caused uninitialized warnings when .value() was
called in the one statement, but it did. Likely due to lifetime rules of
the statement preceding .value()?

Anyway, this makes newer GCC sufficiently happy for the code to compile.

Signed-off-by: Olof Johansson <olofj@tenstorrent.com>