#5480: FD 2.0 prototype prefetcher/dispatcher
The dispatcher test uses a spoofed prefetcher and can write to the worker.
Code is close to final and performance is good, though many more commands need
to be fleshed out.  Testing L1 addresses are hacked for now.

The prefetcher is hacked fairly heavily, does not handle HostQ or CmdDatQ
wrapping, addresses are hacked, etc.  However, it does execute a command
through to the dispatcher.