#0: Speedup incremental builds
This change removed the use of phony targets and also refactored things
in the python_env makefiles to avoid rebuilding the python env for
incremental builds, which takes a significant amount of time.  Largely
this involved ensuring that a file is produced for each target so that
make can track the target's state.