#0: make logLocationsRecord a static function
Global constructors can be visited in any order, when using clang
operationProfiler ends up being constructed before logLocationsRecord so
when referenced in the constructor it's just an empty string.  The fix
is to move logLocationsRecord into a static function that gets lazily
constructed on the fly, even from another global ctor.