#6388: temporarily disable bidirectional support for all-gather
This feature seems to expose a fast dispatch hang. The suspected delta between bidirectional and unidirectional that could expose an issue in the dispatcher path is the extra workers that are instantiated, and thus, the extra binaries that must be sent. The number of workers is doubled when switching to bidirectional.