#7528: add a bidirectional ethernet bandwidth ubench test
Adds a minimal implementation of a multi-channel, bidirectional ethernet
bandwidth benchmark. Best speeds recorded so far (without erisc IRAM
enabled) was 20-21 GBps. This is still noticeably short of the
theoretical max of the link or even the unidirectional bandwidth, when
doubled.

Additional measurements should be made after IRAM support is added