#7921 pyproject: Update a few python dependencies
Update to newer setuptools, numpy and Pillow to make sure Python 3.8-3.12
are supported.

Unforunately, newer numpy is no longer supporting python 3.8, so there's
a need to do a flexible version (>=1.24.4) to allow both recent and old
enough versions.

In case the system has an older setuptools, by default pip doesn't
upgrade it. Add explicit --upgrade to allow that to happen.

Signed-off-by: Olof Johansson <olofj@tenstorrent.com>