#7921 tt_metal/perf_microbenchmark: fixup warnings with newer GCC
GCC 13 found one uninitialized variable warning.

It also complains on the serialization of the CQDispatchCmd struct since it
detects that part of the union might be uninitialized in some code paths.

Seems that this whole construct would be ripe for a C++ constructor
instead, but for now convert it to a memcpy-based temporary buffer to
make the compiler happy.

Signed-off-by: Olof Johansson <olofj@tenstorrent.com>