#0: Add core range iterators
We have CoreRange and CoreRangeSet classes that represent a single rectangle or
a set of rectangles on a 2D grid of cores. There is currently no way to iterate
over instance of these classes, so instead we iterate over all the device cores
and then check whether a core belongs to a core range set. This PR adds the
ability to directly iterate over these classes.