#0: Fix undefined memory bug in `ssm_prefix_scan`
Previously there was a bug where the writer would copy too many bytes
when modifying the previous hidden state in-place. This was because we
were using the number of chunks to determine the size of the copy -
which is incorrect because this value is rounded to the nearest chunk
size (1024).