#7528: add new ethernet microbenchmark, cleanup and re-enable others
Adds new ping latency microbenchmark which measures the time taken (in
cycles) for an erisc to send a message (of user specifyable size) to the
core on the other end of the ethernet link and receive a response.

Other attributes that can be specified are the number of samples to
measure as well as the number of concurrent channels to send pings over
(where each channel is a separate, concurrent transaction/packet on the
same link).

Initial results indicate an almost exactly 1us roundtrip ping time for a
flit (16B payload + 16B message available tail - not a true flit but
very close)

Bidirectional bandwidth and ring latency microbenchmarks are also
re-enabled and improved.

Ring latency improves to about 655ns per hop for an 8 hop ring.

Banwidth is improved to around 23.5GBps for packet size range from 4k to
16k per packet.

Some samples. Buffer size in bytes is on the left. Number of
channels/buffers instantiated per direction are on the right.
So a num channels of 2 indicates 4 channels total (2 in sender direction
and 2 in receiver).

| Buf   |    |
| Size  | #  |
| (B)   | Ch |
|-------|----|
| 36864 | 2  |
| 32768 | 2  |
| 24576 | 3  |
| 16384 | 4  |
| 8192  | 9  |
| 4096  | 18 |
| 2048  | 30 |
| 1024  | 30 |