#10415: add new full-tensor bidirectional mode to all-gather
The new bidirectional all-gather mode is being added as a prerequisite
to all-gather + matmul fusion. In addition, this change also leads to
performance improvements, particularly for smaller all-gathers because
fewer end-to-end latencies add up for what tensor to be single packet
per channel/per ring index.

The new mode sends the full input tensor for a given tensor both
directions around the ring, but only halfway around the ring in each
direction. This is in contrast to the prior default mode (SPLIT_TENSOR)
which would send half of the input tensor each direction, but the full
way around the ring.

This new mode is not enabled yet for sharded all-gather.