#10246: Enable runtime arg double buffering
Only runtime args are double buffered, CB configs and Sems come soon, binaries
down the road

The kernel config buffer is a ring buffer, but the size is presently the same
size as the max number of runtime args.  This means that for large numbers of
runtime args a stall comes before the runtime args are transferred, but in
more common cases the stall comes after the runtime args and before the
other config data.

Note that other stalls for, eg, write buffer, could update the kernel config
buffer so the next enqueue program may proceeed without a stall.  The prior
code didn't handle this case and I didn't add it...todo.