#11323: Enable all UNet tests in CI
This change enables a functional test in the nightly pipeline, as well
as end-to-end and device performance tests in their corresponding
pipelines.

Removed an old performance benchmarking script that is disabled, not
mentioned in any CI pipelines, and no longer works, in favor of the one
in `models/experimental`.

Added some additional asserts to ensure users are running the correct
configuration before using the model.