#11706: Use new Conv2D API in UNet Shallow
A complete rework of this model was required to use the new Conv2D API.
Model preprocessing was simplified and many new unit-level tests were
created to improve maintainability.

The previous version was completely removed. The old demo code (found in
`functional_unet/demo`) was deleted - instead of ported - because it was
non-functional anyway.

Device/end-to-end performance is the same as the old version.