#12103: add contiguous pages in row to width/block/height sharded addrgen
This change allows the caller to amortize calls to the address generator
when multiple contiguous pages are in the row in the bank (shard in case
of sharded address generators), which improves SW performance.

Additionally, by enabling the caller to sequence larger contiguous
chunks of pages, fewer noc commands may be issued - reducing contention
on the noc command buffers, and also inducing lower SW overheads.