#12101: Fixing reduce to account padding while calculating output dimension (#12274)
Problem description
While developing the TT-MLIR compiler, we encountered an issue when using ttnn.mean op. The problem occurs with ttnn.mean operation when input tensor dims aren't tile-dim aligned. The code errors out with the following error message:

Unable to reshape a tensor in TILE_LAYOUT to non-tile height and width! Please convert the tensor to ROW_MAJOR_LAYOUT first.

The issue is in the TTNN implementation of reduction op. The issue is in this line where we calculate the padding for the output shape:
padded_output_shape.push_back(input_shape[axis]);

We iterate through all the dims of the tensor. When we encounter the dim that we are not reducing, we want to capture the non-padded and padded parts of the input shape, but we don't capture the padded part correctly because input_shape[axis] returns the unpadded part. Instead, we should retrieve something like this:
padded_output_shape.push_back(input_shape.value[axis]);

What's changed
The change includes the padding in the calculation of the output tensor. This change impacts all reduction ops, which now work for dimensions not tile-aligned (like tensor(1, 63, 37)).