#12524: Update fmt and unify logging API (#12464)
#0: Update fmt and unify logging API

Our logging macros accept variadic arguments and print them as-is,
requiring callers to use fmt::format to prepare the arguments. But the
low-level log_fatal, log_warning and so on functions accept a runtime
format string.

This PR unifies the logging API by requiring users to always specify a
valid format string, and avoids runtime formatting.

Force-merging by @tt-rkim  because @TT-BrianLiu is out of office. Post-commit should be passing.