#0: Fix ND hang in test_prefetcher w/ -spre -sdis
With split prefetcher/dispatcher the downstream component sets the MSB of the
credit a terminate signal.  The credit balancing wait waited for the credits to
match the expected credits but didn't account for the terminate signal.  This
would result in a hang if the downstream component finished before the upstream
omponent (which should be and was rare).