#0: If clean reset fails, try continuing on to FW init regardless
For cases where running right after a tt-smi reset, there's a chance
that L1 comes up in a random state that tricks metal into thinking
there's kernels left running. In that case it may try to force exit
them, but if a kernel is not actually running this can result in a hang.
In this case it's actually fine to run kernels on the core, so just
continue on and only error out and ask the user to reset if FW init
fails.