#0: update all-gather tests to remove all_devices test fixture
The all_devices test fixture is unreliable for producing expected chip orderes that are needed for certain CCL operations such as ring all-gathers. Without this change, from machine to machine, the chip order may be incorrect from the order expected by all-gather. To correct the chip, we use fixtures that account for chip ordering.

In particular, 4-chip ring tests that are expected to run on the inner 4-chip ring of the t3000 were updated to use the pcie_mesh_device fixture to ensure that the first 4 chips in the list are those devices in the inner ring.

Additional changes along with the above:
- Removed some redundant `run_all_gather` type test functions to merge into a singla test function
- Added additional n300 all-gather test cases
- Improve error messaging