#13541: Conv2d auto shard interleaved tensor
Conv2d requires (internally) sharded tensor for in0(activations).
Currently if shard_layout is not provided as part of the Conv2dConfig,
op will default to height sharding. And op will run an op to convert
input tensor to height sharded tensor.

This isn't always the best heuristic. This change introduces a logic
to select the best sharding layout based on the input tensor shape, in
order to increase core count and consequently increase L1 space op
can use, and reduce number of out-of-memory errors.
There is still no guarantee that op will fit in L1 space, but we can
observe high success rate in tests coming out of pytorch2 ttnn
integration.

This heuristic is not the most optimal one and we will keep iterating on
it as well, but given that it resolve a lot of torch tests cases we want
to prioritize getting it in.

This change is also a BREAKING CHANGE, as it changes the default behavior
of Conv2d when user passes in interlaved tensor for in0(activations).
This doesn't seem to affect our models given that they already provide
sharded tensors as inputs, but it does affect some of our tests.

In case users wants to pass in interleaved tensor, and manually pick
sharding layout, user has to set shard_layout in Conv2dConfig, which
now defaults to std::nullopt instead of HEIGHT_SHARDED.

To test this change a variant with auto shard selection has been added
to most of tests/ttnn/unit_tests/operations/test_new_conv2d.py.
Some tests are skipped (like stable diffusion tests) as this change
has exposed few issues with conv WIDTH_SHARDED codepath that we need to
address.