Adding script that installs tt-debugger into current python environment (#13534)
Adding simple script that installs tt-debugger into current python environment. At the moment it is heavy, until tt-debugger becomes public repository. Idea is to have script to enable people to test it with tt-metal, and later we will make it install faster from public repository releases instead of building it on install.