#13787: Fix build with CMAKE_DISABLE_PRECOMPILE_HEADERS (#13700)
### Ticket
closes #13787 
needed for #13698

### Problem description
When building with PCH disabled (set `CMAKE_DISABLE_PRECOMPILE_HEADERS` to True), the build fails because not all files include the headers they use.  They are only accidentally able to resolve certain dependencies.

This is a problem, because the first cut of `ccache` is going to skip PCH.  Also analysis tools (including the compiler) don't process non-included headers.  As evidenced by the one non-include change in this PR that the compiler is now detecting.

### What's changed
* Build with `CMAKE_DISABLE_PRECOMPILE_HEADERS = true`
* Fix failure
* GOTO step 1

### Checklist
- [x] [Post commit CI passes](https://github.com/tenstorrent/tt-metal/actions/runs/11335836891)
- [ ] Blackhole Post commit (if applicable)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [x] New/Existing tests provide coverage for changes

https://github.com/tenstorrent/tt-metal/actions/runs/11335836891

---------

Co-authored-by: Bryan Wilder Field Lozano <blozano@tenstorrent.com>