#14464: Re-enable split_reader on gs resnet
After commit #0194005 Resnet50 device perf
regressed to 6600.
Commit above decoupled two conv2d features
shallow_conv and split_reader.
Previously if shallow conv was enabled,
split_reader would be enabled under the hood even
if it's not requested.
Resnet50 on wh was configured to use split_reader,
but on gs it was configured not to use split_reader
and this lead to perf regression.

This change configures split_reader on gs explicitly
to restore the perf.