#14501: Stall enqueue_program after enqueue_trace (#14503)
* #14501: Stall enqueue_program after enqueue_trace

enqueue_program normally writes out the RTA to the ringbuffer before it
stalls waiting for the previous program to finish. If a kernel is
running due to enqueue_trace, the config buffer manager is unaware of the
memory that's in use by the kernel, so the RTA may overwrite the kernel
binary or other L1 memory that's in use.

The WorkerConfigBufferMgr should be informed that its state is out of
date, so it will cause a stall before writing out the next data.