#14571: Check CMake compatibility (#14573)
Ticket
#14571

Problem description
We advertise a certain CMake compatibility, but our CI systems have a newer version for convenience. We need a quick check to ensure we don't slip up and put something in that requires a newer version.

What's changed
New check that uses a fixed (old) version of CMake and confirms whether it can generate build files.