#11795: Dispatch async to ringbuffer
Don't wait for previous programs to finish before copying new ones into the
ringbuffer. This allows more parts of dispatch to work in parallel with
programs.

Ethernet kernels aren't placed in the ringbuffer, so any programs including
ethernet kernels dispatch the same as previously.

Also move the WorkerConfigBufferMgr from the SystemMemoryManager to the
HWCommandQueue, since it relies on sync counts that are dispatcher-specific.
Between command-queues we need to wait for the commands to completely drain,
since we have no way to wait for a specific command in a different command
queue to finish.