Increase packer precision for bfp8 formats (#14822)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-metal/issues/14032)

### Problem description
bfp8 packing is inaccurate when pack_src_format is also bpf8, since this
results in double rounding in the HW. First the gasket rounds to 7 bits,
then rounding occurs again when the mantissas are being shifted in order
to have common exponent.

### What's changed
Add a flag to compute config called `bfp_pack_precise` which toggles the
pack_src_format to either fp16 or fp32 (depending on fp32_mode_en) in
order to get more accurate output. This however will half the packer
bandwidth in the case of fp16, and reduce it to one quarter in the case
of fp32.