#14546: Fix moreh_adamw power_tile reduce performance (#14927)
### Ticket
Link to Github Issue:
https://github.com/tenstorrent/tt-metal/issues/14546

### Problem description
The current power_tile implementation has a time complexity that grows
linearly with the increment step. As a result, a larger step value
significantly increases the kernel’s runtime.

```cpp
        // bias_correction1 = 1 - pow(beta1, step);
        // cb_tmp2 = pow(beta1, step);
        tile_regs_acquire();
        cb_reserve_back(cb_tmp2, onetile);
        copy_tile_init_with_dt(cb_scalar_args);
        copy_tile(cb_scalar_args, beta1_tile, dst0);
        power_tile_init();
        power_tile(dst0, step);
        tile_regs_commit();
```

### What's changed
The exponent of beta1 and beta2 is now computed on the host (program
factory) and then passed to the reader kernel. cb_beta1_exponent and
cb_beta2_exponent is the intermed cb for compute kernel to use. For
cached_program, I overwrite the beta1_exponent and beta2_exponent values
in the reader kernel runtime args.

### Checklist
- [x] Post commit CI passes:
https://github.com/tenstorrent/tt-metal/actions/runs/11850545658
- [x] Blackhole Post commit (if applicable): NA
- [x] Model regression CI testing passes (if applicable): NA
- [x] Device performance regression CI testing passes (if applicable):
NA
- [x] New/Existing tests provide coverage for changes: All tests passed