#11795: Temporarily disable async dispatch
We're seeing some hangs with llama, so temporarily disable async dispatch by
reducing launch message queueing to 1, so every command is forced to wait for
the previous command. This will reduce performance to worse than pre-async
(since we'll always wait before sending  RTAs), but will at least allow us to
exercise most of the async codepaths and continue to validate that the sync
counts received from the config buffer manager are correct.