#15167: explicitly check for rank 4 in reduce special cases (#15181)
### Ticket
Link to Github Issue
https://github.com/tenstorrent/tt-metal/issues/15167

### Problem description
A code path in reduce assumed that rank is 4. This was implicit when
negative dimensions were used at that point because the comparison was a
dimension against a rank offset. After a change to adjust dimensions to
be non-negative earlier in the code, an incorrect code path was taken in
some scenarios.

### What's changed
Explicitly check that the rank is 4.

### Checklist
- [x] Post commit CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/11899700948/job/33161083860
vs
https://github.com/tenstorrent/tt-metal/actions/runs/11899483294/job/33160246174
- unrelated issues, has to be good enough
- [ ] Blackhole Post commit (if applicable) N/A
- [ ] Model regression CI testing passes (if applicable) N/A
- [ ] Device performance regression CI testing passes (if applicable)
N/A
- [x] New/Existing tests provide coverage for changes