#14826: reorganize crt startup (#15094)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/14826

### Problem description
the original fix for the issue caused a performance regression and was
reverted. This more incremental approach will help determine which bit
was responsible. This particular diff moves things around

### What's changed
1) The firmware did bss clearing and global initializer running from
assembly startup code, and *then* did initialized data copying. That
only 'works' because we have no global initializers. Remove the assembly
and just do all 3 initializations in the correct order from the
conveniently available `firmware_kernel_common_init`, which doesn't seem
to be used from firmware, just kernels -- 'firmware-kernels'?
2) Rename that routine as do_crt1, because that's what it's doing.
3) This also catches the case fixed by
https://github.com/tenstorrent/tt-metal/pull/15009, where I'd missed a
case.

In particular this patch does NOT
*) insert unroll-inhibiting pragmas
*) change the data memory copy
*) change the bss clearing routine (which the optimizer spots is memset,
and substitutes that, leading to code bloat).

### Checklist
- [YES] Post commit CI passes
- [ ] Blackhole Post commit (if applicable)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [ ] New/Existing tests provide coverage for changes