Add forward support for PReLU (#14940)
### Ticket
Link to Github Issue #8544 

### Problem description
Currently PReLU is aliased to LeakyRelu which is not correct. It needs
to be properly implemented in the eltwise operation.

### What's changed
- Update the prelu implementation In composite structure.
- It also support sfpu(Added in this PR) for certain cases which will
handled soon in #14933 .

### Additional Information
In Torch PReLU, the second input tensor can only have two valid shapes:
either a tensor with a single value (1) or a tensor with a size equal to
the number of input channels (default is 1). Currently, This
implementation only supports cases where it matches the number of
channels. Support for a single-value tensor requires additional handling
at the low-level kernel.

### Checklist
- [ ] [All Post commit CI
](https://github.com/tenstorrent/tt-metal/actions/runs/11796920085)