#0: Dispatch RTAs early in some cases
Often, a full launch message buffer will be the only reason the new program
can't be sent. This is particularly common now with the launch message buffer
being hardcoded to 1 entry (to prevent async dispatch), but it could happen
anyway.

In that case, we could send both the kernel binaries and the RTAs. The code is only set
up to sync before the RTAs or before the kernel binearies, so for now we still have to
wait before sending the binaries, even if they would fit.