Support for new matmul1d op with gather_in0 (#14964)
### Ticket
- #14962

### Problem description
Currently, matmul 1d supports mcast_in0, for when the input is sharded
across all the cores. However in some cases (specifically, the case of
the matmuls used in the Llama models), this poses a bottleneck to the
compute, as each core must wait to receive each shard prior to
processing it.

To combat this, a new option to matmul1d is proposed: `gather_in0`.
Using this, the activations are gathered using a ring all-gather
operation. This allows each core to start processing the local
activation shard that is already available, and process other shards as
soon as they arrive. Essentially, this overlaps the time taken to gather
the activation and to do the computation.

See issue for a diagram of how the inputs are gathered.

### What's changed
This PR adds the following changes:
1. A new `gather_in0` flag, defaulted to false, in the
`MatmulMultiCoreReuseMultiCast1DProgramConfig` header and pybind
2. A new helper function in the matmul1d program factory, to
specifically handle the case when `gather_in0=True`
3. 3 new kernels
  i. a ring gather in0 kernel
  ii. an in1 kerenel
  iii. a bmm kernel (close copy of the existing one)
4. Validation in `matmul_op.cpp` for the new gather_in0 case
5. `test_matmul_1d_gathered.py` to test the new matmul configuration

### Checklist
- [x] Post commit CI passes
([link](https://github.com/tenstorrent/tt-metal/actions/runs/12054246113))
- [x] New/Existing tests provide coverage for changes