Add missing check in mul_tiles_bcast_cols (#15754)
### Ticket
#14352 #14349

### Problem description
Tests from ```misc/test_layernorm.py``` and ```misc/test_softmax.py```
are failing. They are calling mul_tiles_bcast_cols, which is incorrect
for BH for two reasons:
1. There's a missing check when clearing faces 0 and 1. These faces
shouldn't be cleared from DEST unconditionally, but only if DEST result
is reused.
2. ZEROACC instructions in FP32 case are being incorrectly used for BH.
That is why despite the first reason, the tests pass on WH, because
there two ZEROACC instructions are needed, each of which clears half of
face. Since this must be tested in more detail, it's moved to be fixed
as a part of #15713.

### What's changed
This PR does the following:
1. Adds the missing DEST reuse check.
2. Enables the tests from ```misc/test_layernorm.py``` and
```misc/test_softmax.py```.
3. Since it pulls the changes from `tt-llk-bh`, it removes the assert
for reduce scalar in fp32 mode, thus enabling #15744


### Checklist
- [ ] Post commit CI passes - [not
necessary](https://github.com/tenstorrent/tt-metal/actions/runs/12185815043)
- [x] Blackhole Post commit (if applicable) -
[#2127](https://github.com/tenstorrent/tt-metal/actions/runs/12185808879)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [x] New/Existing tests provide coverage for changes