#15747: Giving bfloat8 support to reshape (#15751)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-metal/issues/15747)

### Problem description
I neglected to add support for bfloat8_b as none of the tests looked for
it. I am adding both the support and an extra test

### What's changed
Added bfloat8_b support. As per the docs, bfloat8_b will only work on
unpadded tiles, so this is what we support and it asserts out in other
cases.
### Checklist
- [x] Post commit Passes
https://github.com/tenstorrent/tt-metal/actions/runs/12188208619
- [ ] Blackhole Post commit (if applicable)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [ ] New/Existing tests provide coverage for changes